Small restructuring to make it easier to add additional components
