+++ src/index.scss
@@ -40,7 +40,7 @@
     background-color: rgba(0, 0, 0, 0.85) !important;
     transition: .3s ease !important;
     transition-property: background-color, color !important;
-    
+
     &:hover {
         background-color: rgba(13, 92, 157, 0.85) !important;
         color: inherit !important;
